fix(auth): block requests after Redis session expiry, align initial session TTL
- authenticate() теперь возвращает 401 SESSION_EXPIRED когда Redis доступен,
  но ключ сессии отсутствует (истёк по TTL). Ранее fallback пропускал запрос
  через JWT, игнорируя session_lifetime_minutes.
- Деградация на JWT сохранена только при недоступном Redis.
- Системные аккаунты (@flow-universe.internal) освобождены от проверки сессии.
- getSessionLifetimeMinutes вынесена в shared/utils/session-settings.ts.
- setUserSession принимает опциональный ttlSeconds; login/register/refresh
  передают session_lifetime_minutes * 60 вместо хардкода 7 дней.
