+++ backend/src/shared/middleware/auth.ts
@@ -3,31 +3,11 @@ import type { SystemRoleType } from '@prisma/client';
 import { verifyAccessToken } from '../utils/jwt.js';
 import { AppError } from './error-handler.js';
 import type { AuthRequest } from '../types/index.js';
-import { getUserSession, touchUserSession, getCachedJson, setCachedJson, isRedisAvailable } from '../redis.js';
-import { prisma } from '../../prisma/client.js';
-
-const DEFAULT_SESSION_LIFETIME_MINUTES = 60;
+import { getUserSession, touchUserSession, isRedisAvailable } from '../redis.js';
+import { getSessionLifetimeMinutes } from '../utils/session-settings.js';
 
 // In-process counter — exported so health/metrics endpoints can expose it.
 export const sessionFallbackCounter = { total: 0, redis_unavailable: 0, session_missing: 0 };
-const SETTING_CACHE_TTL_SECONDS = 60;
-const SESSION_LIFETIME_SETTING_KEY = 'session_lifetime_minutes';
-const SESSION_LIFETIME_CACHE_KEY = `settings:${SESSION_LIFETIME_SETTING_KEY}`;
-
-async function getSessionLifetimeMinutes(): Promise<number> {
-  const cached = await getCachedJson<number>(SESSION_LIFETIME_CACHE_KEY);
-  if (cached !== null) return cached;
-
-  try {
-    const setting = await prisma.systemSetting.findUnique({ where: { key: SESSION_LIFETIME_SETTING_KEY } });
-    const value = setting ? parseInt(setting.value, 10) : DEFAULT_SESSION_LIFETIME_MINUTES;
-    const result = isNaN(value) || value < 5 ? DEFAULT_SESSION_LIFETIME_MINUTES : value;
-    await setCachedJson(SESSION_LIFETIME_CACHE_KEY, result, SETTING_CACHE_TTL_SECONDS);
-    return result;
-  } catch {
-    return DEFAULT_SESSION_LIFETIME_MINUTES;
-  }
-}
 
 export async function authenticate(req: AuthRequest, _res: Response, next: NextFunction) {
   const header = req.headers.authorization;
@@ -74,16 +54,26 @@ export async function authenticate(req: AuthRequest, _res: Response, next: NextF
       return next(new AppError(401, 'Session expired due to inactivity', { code: 'SESSION_EXPIRED' }));
     }
   } else {
-    // session===null has three causes: (1) Redis unavailable, (2) session key expired/missing,
-    // (3) session was never created (system accounts). Degrade gracefully — rely on JWT expiry.
-    const reason = (await isRedisAvailable()) ? 'session_missing' : 'redis_unavailable';
-    sessionFallbackCounter.total += 1;
-    sessionFallbackCounter[reason] += 1;
-    console.warn('[auth] sliding-session fallback', {
-      userId: payload.userId,
-      reason,
-      counter: sessionFallbackCounter,
-    });
+    // session===null: (1) Redis unavailable, (2) session key expired, (3) system accounts.
+    if (await isRedisAvailable()) {
+      // Redis is up — session key is gone, meaning it expired or was never created.
+      // System accounts (agent@) never get a session key; allow them through.
+      // Regular users with an expired key must re-authenticate.
+      const isSystemAccount = payload.email.endsWith('@flow-universe.internal');
+      if (!isSystemAccount) {
+        sessionFallbackCounter.total += 1;
+        sessionFallbackCounter['session_missing'] += 1;
+        return next(new AppError(401, 'Session expired due to inactivity', { code: 'SESSION_EXPIRED' }));
+      }
+    } else {
+      // Redis is unavailable — degrade gracefully, rely on JWT expiry.
+      sessionFallbackCounter.total += 1;
+      sessionFallbackCounter['redis_unavailable'] += 1;
+      console.warn('[auth] sliding-session fallback: redis unavailable', {
+        userId: payload.userId,
+        counter: sessionFallbackCounter,
+      });
+    }
   }
 
   next();
